+++ src/main/java/com/global/api/entities/enums/ApplicationCryptogramType.java
@@ -2,5 +2,6 @@ package com.global.api.entities.enums;
 
 public enum ApplicationCryptogramType {
     TC,
-    ARQC
+    ARQC,
+    AAC
 }